+++ addon/components/flexberry-objectlistview.js
@@ -155,6 +155,15 @@ export default FlexberryBaseComponent.extend({
   editFormRoute: undefined,
 
   /**
+    Flag indicates whether component on edit form (for FOLV).
+
+    @property onEditForm
+    @type Boolean
+    @default false
+  */
+  onEditForm: false,
+
+  /**
     Primary action for row click.
 
     @property action
@@ -485,12 +494,6 @@ export default FlexberryBaseComponent.extend({
   */
   recordsTotalCount: null,
 
-  perPageValueObserver: Ember.observer('perPageValue', function() {
-    let perPageValue = this.get('perPageValue');
-
-    this.get('userSettingsService').setCurrentPerPage(this.componentName, undefined, perPageValue);
-  }),
-
   /**
     Current interval of records.
 
@@ -623,9 +626,10 @@ export default FlexberryBaseComponent.extend({
 
       @method actions.objectListViewRowClick
       @public
-      @param {Object} record Clicked record
+      @param {Object} record Clicked record.
+      @param {Object} options Different parameters to handle action.
     */
-    objectListViewRowClick(record) {
+    objectListViewRowClick(record, options) {
       let $clickedRow = this._getRowByKey(record.key || Ember.guidFor(record));
       Ember.run.after(this, () => { return $clickedRow.hasClass('active'); }, () => {
         if (this.get('componentMode') === 'lookupform') {
@@ -633,7 +637,14 @@ export default FlexberryBaseComponent.extend({
         } else {
           let editFormRoute = this.get('editFormRoute');
           Ember.assert('Edit form route must be defined for flexberry-objectlistview', editFormRoute);
-          this.sendAction('action', record, editFormRoute);
+          if (Ember.isNone(options)) {
+            options = {};
+            options.editFormRoute = editFormRoute;
+          } else {
+            options = Ember.merge(options, { editFormRoute: editFormRoute });
+          }
+
+          this.sendAction('action', record, options);
         }
       });
 
@@ -859,9 +870,14 @@ export default FlexberryBaseComponent.extend({
       Called when click on perPage.
 
       @method actions.perPageClick
+      @param {String} perPageValue Selected perPage value.
     */
-    perPageClick() {
-      this.get('eventsBus').trigger('showLoadingTbodyClass', this.get('componentName'), true);
+    perPageClick(perPageValue) {
+      var userSettings = this.get('userSettingsService');
+      if (parseInt(perPageValue, 10) !== userSettings.getCurrentPerPage(this.componentName)) {
+        userSettings.setCurrentPerPage(this.componentName, undefined, perPageValue);
+        this.get('eventsBus').trigger('showLoadingTbodyClass', this.get('componentName'), true);
+      }
     },
 
     /**