+++ tests/dummy/app/locales/en/translations.js
@@ -605,7 +605,7 @@ Ember.$.extend(true, translations, {
           'custom-button-name': 'Send hello'
         },
         'on-edit-form': {
-          'caption': 'Flexberry-objectlistview. FlexberryObjectlistview custom data sample'
+          'caption': 'Flexberry-objectlistview. FlexberryObjectlistview on edit form'
         },
         'custom-filter': {
           'caption': 'Flexberry-objectlistview. Custom filter'